requirements.txt
requests
pandas
boto3
python-dotenv
snowflake-snowpark-python
snowflake-cli-labs
